lasted file
lasted file

+++ src/components/App/index.js
@@ -1,7 +1,9 @@
-import React, { Component ,Fragment } from 'react';
-import { Table ,Container, Modal ,ModalHeader ,ModalBody,ModalFooter} from 'reactstrap'
+import React, { Component } from 'react';
+import { connect } from 'react-redux'
+import {Button , Table ,Container, Modal ,ModalHeader ,ModalBody,ModalFooter} from 'reactstrap'
+import { setMovieInfo } from '../../modules/ListMovie/actions'
 import Main from './style'
-import { Button,Dropdown,Form } from 'react-bootstrap';
+import { Dropdown,Form } from 'react-bootstrap';
 import {
   Textbox,
   Select
@@ -15,34 +17,35 @@ class App extends Component {
       modal: false,
       newMovie: "",
       list: [],
-      flag:'create',
-      trick:false
+      flag:'create', 
+      thumb:[],
+      filterStatus : "",
+      filterCategory : "",
+      cSelected:[],
+      hasImgError : true,
+      hasNameError : true,
+      hasCategoryError : true,
+      hasStatusError : true,
+      validate : false
     };
-    this.validateForm = this.validateForm.bind(this);
   }
+
   toggleValidating(validate) {
     this.setState({ validate });
   }
 
-  toggle = this.toggle.bind(this);
-  validateForm(e) {
-    e.preventDefault();
-    this.toggleValidating(true);
-    const { 
-      hasImgError,
-      hasNameError,
-      hasCategoryError,
-      hasStatusError
-    } = this.state;
-    if (
-      !hasImgError &&
-      !hasNameError &&
-      !hasCategoryError &&
-      !hasStatusError
-    ) {
-      alert("All validated!");
+  onCheckboxBtnClick(selected) {
+    const index = this.state.cSelected.indexOf(selected);
+    if (index < 0) {
+      this.state.cSelected.push(selected);
+    } else {
+      this.state.cSelected.splice(index, 1);
     }
+    this.setState({ cSelected: [...this.state.cSelected] });
   }
+
+  toggle = this.toggle.bind(this);
+ 
   toggle() {
     this.setState(prevState => ({
       modal: !prevState.modal,
@@ -54,38 +57,34 @@ class App extends Component {
         name:"",
         category:"",
         status:""
-    }));
+    }))
   }
 
   updateInput(key, value) {
     this.setState({ [key]: value });
   }
 
-
   componentDidMount(){
     const storeMovie = localStorage.getItem('store');
     if(storeMovie !== null){ 
       this.setState({list:JSON.parse(storeMovie)})
+      this.setState({thumb:JSON.parse(storeMovie)})
+      setMovieInfo(JSON.parse(storeMovie))
     }
   }
 
-  addMovie() {
+  addMovie = () => {
     this.toggleValidating(true);
-    const { flag ,list,index} = this.state
-    
-    if(flag==='create'){
-      const { 
+    const { flag,list,index} = this.state
+    if(flag === 'create'){
+
+      let { 
         hasImgError,
         hasNameError,
         hasCategoryError,
         hasStatusError
-      } = this.state;
-      if (
-        !hasImgError &&
-        !hasNameError &&
-        !hasCategoryError &&
-        !hasStatusError
-      ) {
+      } = this.state
+      if (!hasImgError && !hasNameError && !hasCategoryError && !hasStatusError ) {
         const { img,name,category,status } = this.state
         const newMovies = {
           id: 1 + Math.random(),
@@ -93,58 +92,69 @@ class App extends Component {
           name,
           category,
           status
-        };
-  
+        }
         const list = [...this.state.list];
-  
         list.push(newMovies);
         localStorage.setItem('store',JSON.stringify(list))
-  
+        setMovieInfo(JSON.stringify(list))
         this.setState({
           list,
           newMovie: "",
           img:"",
           name:"",
           category:"",
-          status:"",
-  
+          status:""
         });
-  
         this.toggle()
-      }
+      } 
     }else{
-      console.log('state : ',this.state)
-      this.setState({trick:true})
-
+      alert('Update data success')
       list[index].img = this.state.img
       list[index].name = this.state.name
       list[index].category = this.state.category
       list[index].status = this.state.status
-      console.log('list : ',list)
       localStorage.setItem('store', JSON.stringify(list))
-      // this.setState({img : 'xxx'})
-      // this.state.img = this.state.list[index].img
-      // this.setState(this.state.img)
-      // this.setState({})
-      // this.setState({list[index] img });
-       
+      this.toggle()
+    
     }
-    
-    
   }
 
-  deleteItem(id) {
-    const list = [...this.state.list]
-    const storeMovie = localStorage.getItem('store')
+  deleteItem = (id,flagDel = '') => {
+    if(flagDel == 'all'){
+      const list = [...this.state.list]
+      const storeMovie = localStorage.getItem('store')
+      let movieListStore = JSON.parse(storeMovie)
+      const updatedListStore = movieListStore.filter(item => item.id !== id)
+      const updatedList = list.filter(item => item.id !== id)
+      this.setState({ list: updatedList })
+      this.setState({ thumb: updatedList })
+      this.setState({ list: updatedListStore })
+      localStorage.setItem('store', JSON.stringify(updatedListStore))
+    }else{
+      if (window.confirm('Are you sure to delete this item?')){
+        const list = [...this.state.list]
+        const storeMovie = localStorage.getItem('store')
+        let movieListStore = JSON.parse(storeMovie)
+        const updatedListStore = movieListStore.filter(item => item.id !== id)
+        const updatedList = list.filter(item => item.id !== id)
+        this.setState({ list: updatedList })
+        this.setState({ thumb: updatedList })
+        this.setState({ list: updatedListStore })
+        localStorage.setItem('store', JSON.stringify(updatedListStore))
+      }
+    }
+  }
 
-    let movieListStore = JSON.parse(storeMovie)
-    const updatedListStore = movieListStore.filter(item => item.id !== id)
-    const updatedList = list.filter(item => item.id !== id)
-    this.setState({ list: updatedList })
-    this.setState({ list: updatedListStore })
-
-    localStorage.setItem('store', JSON.stringify(updatedListStore))
-    
+  deleteOptions = () => {
+    const { cSelected } = this.state
+    if(cSelected.length > 0){
+      if (window.confirm('Are you sure to delete this item?')){
+        for(var i=0; i<cSelected.length; i++){
+          this.deleteItem(cSelected[i],'all')
+        }
+      }
+    }
+   
   }
 
   viewItem = (id,index) => {
@@ -161,37 +171,61 @@ class App extends Component {
     }))
   }
 
+  filterBy = (type,action) => {
+    const { thumb } = this.state
+    this.setState({list:thumb})
+      if(action !== 'All'){ 
+        if(type === 'status'){
+          this.setState({list:thumb.filter(item => item.status === action )})
+        }else{
+          this.setState({list:thumb.filter(item => item.category === action )})
+        }
+      }else{
+        this.setState({list:thumb})
+      }
+  }
+
   render() {
     const { 
       img,
       name,
       category,
       status,
-      validate,
-      trick
+      validate
     } = this.state;
-    console.log(trick)
     return (
-      <Container>
-        <Main>
+        <Container>
+          <Main>
           <div className="left">
-              <Button onClick={this.toggle}>Add +</Button>
+              <Button onClick={this.toggle}>Add +</Button> 
           </div>
           <div className="right">
-            <Dropdown>
+            <Dropdown className="fix-position">
               <Dropdown.Toggle variant="success" id="dropdown-basic">
-                Dropdown Button
+                By Category
               </Dropdown.Toggle>
               <Dropdown.Menu>
-              <Dropdown.Item href="#/action-1">Action</Dropdown.Item>
-                <Dropdown.Item href="#/action-2">Another action</Dropdown.Item>
-                <Dropdown.Item href="#/action-3">Something else</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('category','All')}>All</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('category','Action')}>Action</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('category','Drama')}>Drama</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('category','Sci-fi')}>Sci-fi</Dropdown.Item>
+              </Dropdown.Menu>
+            </Dropdown> 
+            <Dropdown className="fix-position">
+              <Dropdown.Toggle variant="success" id="dropdown-basic">
+                By Status
+              </Dropdown.Toggle>
+              <Dropdown.Menu>
+                <Dropdown.Item onClick={() => this.filterBy('status','All')}>All</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('status','Active')}>Active</Dropdown.Item>
+                <Dropdown.Item onClick={() => this.filterBy('status','InActive')}>InActive</Dropdown.Item>
               </Dropdown.Menu>
             </Dropdown> 
           </div>
               <Table responsive striped bordered hover>
                 <thead>
                   <tr>
+                    <th>options</th>
                     <th>#</th>
                     <th>Image</th>
                     <th>Name</th>
@@ -204,6 +238,7 @@ class App extends Component {
             {this.state.list.map((item, index) => {
               return (
                   <tr key={index} >
+                    <td><Button outline color="warning" onClick={() => this.onCheckboxBtnClick(item.id)} active={this.state.cSelected.includes(item.id)}>Select</Button></td>
                     <td>{index+1}</td>  
                     <td>
                       <img src={item.img} alt={item.name}/>
@@ -211,13 +246,16 @@ class App extends Component {
                     <td>{item.name}</td>
                     <td>{item.category}</td>
                     <td>{item.status}</td>
-                    <td> <Button onClick={() => this.viewItem(item.id,index)}>View</Button> <Button variant="danger" onClick={() => this.deleteItem(item.id)}>Del</Button> </td>
+                    <td> <Button color="info" onClick={() => this.viewItem(item.id,index)}>View</Button> <Button color="danger" onClick={() => this.deleteItem(item.id)}>Del</Button> </td>
                   </tr>
               );
             })}
               </tbody>
            
               </Table>
+
+              <Button color="danger" onClick={this.deleteOptions}>Del All -</Button>
+
 
               <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
                 <ModalHeader toggle={this.toggle}>Add Movie</ModalHeader>
@@ -235,7 +273,7 @@ class App extends Component {
                         placeholder="ex : http://"
                         validate={validate} 
                         validationCallback={res =>
-                          this.setState({ hasNameError: res, validate: false })
+                          this.setState({ hasImgError: res, validate: false })
                         } 
                         onChange={(img, e) => {
                           this.setState({ img });
@@ -346,18 +384,23 @@ class App extends Component {
                 </Form>
                 </ModalBody>
                 <ModalFooter>
-                  <Button variant={(this.state.flag === 'update')?'warning':'success'} onClick={() => this.addMovie()} >
+                  <Button color={(this.state.flag === 'update')?'warning':'success'} onClick={() => this.addMovie()} >
                     {(this.state.flag === 'update')? "Update" : "Add"}
                   </Button>{' '}
-                  <Button variant="secondary" onClick={this.toggle}>Cancel</Button>
+                  <Button color="secondary" onClick={this.toggle}>Cancel</Button>
                 </ModalFooter>
               </Modal>
 
             </Main>
-          </Container>
-
+        </Container>
     )
   }
 }
 
-export default App;
+const mapStateToProps = state => ({
+  list: state.list
+})
+
+
+
+export default connect(mapStateToProps)(App)
